Linux: Use the proper Ruby rpath
This fixes a major issue where Ruby would get a bogus Rpath value (see
https://github.com/mitchellh/vagrant/issues/850#issuecomment-6263739).
This bogus value would cause compiled extensions to have a bogus rpath
value (see
https://github.com/mitchellh/vagrant/issues/850#issuecomment-5541549).
This bogus rpath value would cause the extensions to not load properly.

This patch uses LD_RUN_PATH, which is an environmental variable that
absolutely sets the rpath of the binary it is compiling. LD_RUN_PATH is
set absolutely to only the ${ORIGIN}/../lib, which is the correct value
and gets rid of the bogus values.