Mac: Convert to new toolchain to sign packages.
This massive change removes PackageMaker as a dependency for building
Mac OS X packages and uses the new toolchain starting with 10.6:
pkgbuild, productbuild, and productsign.

pkgbuild is used to build the core package that is extracted for
installation. This is basically just the chroot environment that is then
extracted into "/Applications/Vagrant" by the installer. This package is
signed with my developer certificate (on my computer) for Gatekeeper in
Mountain Lion.

productbuild is then used after pkgbuild to build the user-facing
installer. The installer is defined in a "distribution definition" file
which is an XML file. This file defines all the steps and configuration
options for the installer, including a pointer to the package we built
with pkgbuild.

The final installer is also now signed with my developer certificate so
that the installer works even on systems with Gatekeeper enabled (on by
default in Mountain Lion).