Prevent vagrant-share from installing any dependencies
The dependency vagrant-share has is contained within vagrant
proper. Preventing the vagrant-share install from pulling any
dependencies means it will have to resolve its dependencies from
the locally installed set. If that is not possible, it will fail,
which is the correct result to prevent shipping multiple versions
of a single library.