Introduce versioned gem folder per version of Vagrant
Prior to this commit, when users upgraded, they would often run into gem
dependency conflicts. This was because upgrading would leave behind
previously used versions of libraries. This commit updates the installer
to split out the gem installation to go into a version specific folder
based on which version of Vagrant is being installed.