Add stepped packaging scripts for darwin platform
Generation of macOS packages is split into separate steps, the results
of which can be easily cached for re-use. Each step provides an either
and artifact which can be signed (.pkg / .dmg) or the contents of which
can be signed (.zip). This allows the installer DMG to be built in
stages and signed as required.