Remove gem fixes and add extension fixups
Fixes to the Ruby build allow google-protobuf and grpc gems to build
properly using the ruby platform value. Retain the grcp gem cleanup as
the intermediate build files still remain and need to be removed.

RubyGems relies on a touched file to determine if extensions have been
built. This path includes the platform so it must be duplicated for the
non-build host platform to ensure gems will behave properly on both
platforms.