Set permissions explicitly on launcher before packaging it
While `go build` seems to output a binary with the proper permissions,
something about ferrying the launcher binaries through a zip file in GitHub
actions causes it to lose its executable bit.

So here we `chmod` our way to victory just as we arrange the launcher into
place.

Fixes #10
Fixes #11