Update appimage build to use new GitHub Actions CI
Appimages are supposed to be built on the oldest possible operating
system for compatibility. They recommend the oldest supported Ubuntu
LTS.

Previously we accomplished this by building inside a Vagrant box,
but this isn't viable on GitHub Actions runners. Instead, we use the
same chroot strategy we use to build substrates to get an Ubuntu Trusty
environment.