Windows script to build installed gems artifact
This script uses a substate and vagrant gem to install vagrant into the
substrate and then create an artifact of the final installed gems. The
result can be used for signing and reassembling with the substrate for
final packaging.

The grpc script is a helper for handling the grpc rubygem which at the
time of this commit was not shipping a prebuilt gem for the ruby version
being used. It modifies the extconf.rb file prior to extension build to
fix an issue preventing the extension from building. It also will locate
a matching rubygem version for the host (mingw64/mingw32) and download
it to extract the prebuilt grpc core library as local compilation would
error due to max argument length restrictions.