Extract substrate build to reusable workflow
The AppImage build utilizes the substrate built for `.deb` packages.
This extracts that process into an isolated workflow that both can use.

The concurrency added only allows one instance of the workflow to run at
a given time. This is because we can't access the `needs` context from
concurrency so cannot use the generated id on the build job which would
be ideal.