Variable name, build arch, and exist check
Fix typo in variable name. When building launchers for macos packages,
use correct architecture for x86 build. In the debian 64 substrate
workflow, include the info job to run a check for the substrate to
properly skip building when more than one are triggered in a workflow.

Fix variable name and build correct arch