Modify gcc standard on vagrant install
Detect if gcc supports the c23 standard and set the standard to
c17 in the CFLAGS if detected. This will prevent modifying the
standard if it is not required. Perform the same detection in
the substrate build when building libgmp and ruby. For libgmp
it prevents errors on configuration. For Ruby, it prevents issues
with later gem installs that include extensions.